requirements.txt
asgiref==3.3.1
boto3==1.16.36
botocore==1.19.36
dj-database-url==0.5.0
Django==3.1.4
django-ckeditor==6.0.0
django-crispy-forms==1.10.0
django-heroku==0.3.1
django-js-asset==1.2.2
django-location-field==2.1.0
django-model-utils==4.1.1
django-on-heroku==1.1.2
django-storages==1.10.1
gunicorn==20.0.4
jmespath==0.10.0
Pillow==8.0.1
psycopg2==2.7.5
psycopg2-binary==2.8.6
python-dateutil==2.8.1
pytz==2020.4
s3transfer==0.3.3
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
whitenoise==5.2.0
